+++ src/main/java/by/fakeinstagram/web/servlet/DeletePostServlet.java
@@ -1,0 +1,26 @@
+package by.fakeinstagram.web.servlet;
+
+import by.fakeinstagram.service.PostService;
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+
+@WebServlet(urlPatterns = "/deletepost", name = "DeletePostServlet")
+public class DeletePostServlet extends HttpServlet {
+    private final PostService postService = new PostService();
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        long postId = Long.parseLong(req.getParameter("id"));
+        postService.deletePost(postId);
+        resp.sendRedirect("/feed");
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+    }
+}
