+++ src/main/java/by/fakeinstagram/web/servlet/EditPostServlet.java
@@ -1,8 +1,6 @@
-package by.fakeinstagram.servlet;
-
-import by.fakeinstagram.dao.PostDao;
+package by.fakeinstagram.web.servlet;
 import by.fakeinstagram.entity.Post;
-import by.fakeinstagram.entity.User;
+
 import by.fakeinstagram.service.PostService;
 
 import javax.servlet.ServletException;
@@ -12,20 +10,26 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(urlPatterns = "/post", name = "PostServlet")
-public class PostServlet extends HttpServlet {
+@WebServlet(urlPatterns = "/editpost", name = "EditPostServlet")
+public class EditPostServlet extends HttpServlet {
+
     private final PostService postService = new PostService();
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-
-        Post post = (Post) req.getSession().getAttribute("post");
-        getServletContext().getRequestDispatcher("/pages/post.jsp").forward(req, resp);
-        req.setAttribute("post", postService.getPostById(post.getId()));
+        req.setAttribute("post", postService.getPostById(Long.parseLong(req.getParameter("id"))).get());
+        getServletContext().getRequestDispatcher("/pages/editPost.jsp").forward(req, resp);
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        long id = Long.parseLong(req.getParameter("id"));
 
+        Post post = postService.getPostById(id).get();
+        post.setTitle(req.getParameter("title"));
+        post.setDescription(req.getParameter("description"));
+        postService.updatePost(post);
+
+        resp.sendRedirect("/feed");
     }
 }
