+++ src/main/java/by/fakeinstagram/web/servlet/PostServlet.java
@@ -1,6 +1,7 @@
-package by.fakeinstagram.servlet;
+package by.fakeinstagram.web.servlet;
 
 import by.fakeinstagram.service.PostService;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -8,20 +9,19 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-
-@WebServlet(urlPatterns = "/deletepost", name = "DeletePostServlet")
-public class DeletePostServlet extends HttpServlet {
+@WebServlet(urlPatterns = "/post", name = "PostServlet")
+public class PostServlet extends HttpServlet {
     private final PostService postService = new PostService();
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        getServletContext().getRequestDispatcher("/pages/deletePost.jsp").forward(req, resp);
+        long id  =  Long.parseLong(req.getParameter("id"));
+        req.setAttribute("post", postService.getPostById(id).get());
+        getServletContext().getRequestDispatcher("/pages/post.jsp").forward(req, resp);
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        long postId = Long.parseLong(req.getParameter("id"));
-        postService.deletePost(postId);
-        getServletContext().getRequestDispatcher("/pages/feed.jsp").forward(req, resp);
+
     }
 }
