+++ src/main/webapp/pages/feed.jsp
@@ -7,10 +7,9 @@
 </head>
 <body>
 <p>allPosts</p>
-<c:forEach var="postList" items="${requestScope.allPosts}" >
+<c:forEach var="postList"  items="${requestScope.allPosts}" >
     <div>
-        <a href="/post"><c:out value="${postList.title}"/></a>
-
+        <a href="/post?id=${postList.getId()}"><c:out value="${postList.title}"/></a>
         <c:out value="${postList.user.userName}   ${postList.description}    ${postList.dateOfCreation}"/>
     </div>
 </c:forEach>
