+++ src/main/webapp/pages/post.jsp
@@ -1,4 +1,5 @@
 <%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <html>
 <head>
     <title>Post</title>
@@ -7,8 +8,15 @@
 <p>Post</p>
 <form action="/post" method="get">
     <div>
-        <p>${requestScope.post}</p>
+        <p>${requestScope.post.user.userName} ${requestScope.post.title} ${requestScope.post.description} ${requestScope.post.dateOfCreation}</p>
     </div>
 </form>
+
+<!-- валидация по айди что бы кнопка показывалась только на постах текущего юзера-->
+<a href="/editpost?id=${requestScope.post.id}">edit this post</a>
+
+<a href="/deletepost?id=${requestScope.post.id}">delete this post</a>
+
+<a href="/feed">return to feed</a>
 </body>
 </html>
